+++ pkg/providers/property/resource_akamai_property.go
@@ -255,9 +255,11 @@ func resourceProperty() *schema.Resource {
 				Elem:     papiError(),
 			},
 			"rule_warnings": {
-				Type:     schema.TypeList,
-				Computed: true,
-				Elem:     papiError(),
+				Type:       schema.TypeList,
+				Optional:   true,
+				Computed:   true,
+				Elem:       papiError(),
+				Deprecated: "Rule warnings will not be set in state anymore",
 			},
 
 			// Hard-deprecated attributes: These are effectively removed, but we wanted to refer users to the upgrade guide
@@ -504,12 +506,35 @@ func resourcePropertyRead(ctx context.Context, d *schema.ResourceData, m interfa
 	if err != nil {
 		return diag.FromErr(err)
 	}
+	if len(RuleErrors) > 0 {
+		if err := d.Set("rule_errors", papiErrorsToList(RuleErrors)); err != nil {
+			return diag.FromErr(fmt.Errorf("%w: %s", tools.ErrValueSet, err.Error()))
+		}
+		msg, err := json.MarshalIndent(papiErrorsToList(RuleErrors), "", "\t")
+		if err != nil {
+			return diag.FromErr(fmt.Errorf("error marshaling API error: %s", err))
+		}
+		logger.Errorf("Property has rule errors %s", msg)
+	}
+	if len(RuleWarnings) > 0 {
+		msg, err := json.MarshalIndent(papiErrorsToList(RuleWarnings), "", "\t")
+		if err != nil {
+			return diag.FromErr(fmt.Errorf("error marshaling API warnings: %s", err))
+		}
+		logger.Warnf("Property has rule warnings %s", msg)
+	}
 
 	RulesJSON, err := json.Marshal(Rules)
 	if err != nil {
 		logger.WithError(err).Error("could not render rules as JSON")
 		return diag.Errorf("received rules that could not be rendered to JSON: %s", err)
 	}
+	PropertyVersion := Property.LatestVersion
+	res, err := fetchPropertyVersion(ctx, client, PropertyID, GroupID, ContractID, PropertyVersion)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+	Property.ProductID = res.Version.ProductID
 
 	attrs := map[string]interface{}{
 		"name":               Property.PropertyName,
@@ -524,7 +549,6 @@ func resourcePropertyRead(ctx context.Context, d *schema.ResourceData, m interfa
 		"rules":              string(RulesJSON),
 		"rule_format":        RuleFormat,
 		"rule_errors":        papiErrorsToList(RuleErrors),
-		"rule_warnings":      papiErrorsToList(RuleWarnings),
 	}
 	if Property.ProductID != "" {
 		attrs["product_id"] = Property.ProductID
@@ -601,13 +625,14 @@ func resourcePropertyUpdate(ctx context.Context, d *schema.ResourceData, m inter
 		ProductionVersion: ProductionVersion,
 	}
 
-	// load status for what we currently have as latest version.  GetLatestVersion may also work here.
-	resp, err := client.GetPropertyVersion(ctx, papi.GetPropertyVersionRequest{
-		PropertyID:      d.Id(),
-		PropertyVersion: d.Get("latest_version").(int),
-		ContractID:      d.Get("contract_id").(string),
-		GroupID:         d.Get("group_id").(string),
-	})
+	// Schema guarantees group_id, and contract_id are strings
+	PropertyID := d.Id()
+	ContractID := d.Get("contract_id").(string)
+	GroupID := d.Get("group_id").(string)
+	PropertyVersion := Property.LatestVersion
+
+	resp, err := fetchPropertyVersion(ctx, client, PropertyID, GroupID, ContractID, PropertyVersion)
+
 	if err != nil {
 		d.Partial(true)
 		return diag.FromErr(err)
@@ -818,6 +843,27 @@ func fetchProperty(ctx context.Context, client papi.PAPI, PropertyID, GroupID, C
 
 	logger.Debug("property fetched")
 	return res.Property, nil
+}
+
+// load status for what we currently have as latest version.  GetLatestVersion may also work here.
+func fetchPropertyVersion(ctx context.Context, client papi.PAPI, PropertyID, GroupID, ContractID string, PropertyVersion int) (*papi.GetPropertyVersionsResponse, error) {
+	req := papi.GetPropertyVersionRequest{
+		PropertyID:      PropertyID,
+		ContractID:      ContractID,
+		GroupID:         GroupID,
+		PropertyVersion: PropertyVersion,
+	}
+	logger := log.FromContext(ctx).WithFields(logFields(req))
+	logger.Debug("fetching property version")
+
+	res, err := client.GetPropertyVersion(ctx, req)
+	if err != nil {
+		logger.WithError(err).Error("could not read property version")
+		return nil, err
+	}
+	logger = logger.WithFields(logFields(*res))
+	logger.Debug("property version fetched")
+	return res, err
 }
 
 // Fetch hostnames for latest version of given property