+++ pkg/providers/property/resource_akamai_property_test.go
@@ -158,6 +158,7 @@ func TestResProperty(t *testing.T) {
 			resource.TestCheckResourceAttr("akamai_property.test", "group", "grp_0"),
 			resource.TestCheckResourceAttr("akamai_property.test", "product", "prd_0"),
 			resource.TestCheckResourceAttr("akamai_property.test", "product_id", "prd_0"),
+			resource.TestCheckNoResourceAttr("akamai_property.test", "rule_warnings"),
 			resource.TestCheckResourceAttr("akamai_property.test", "rules", `{"rules":{"name":"default","options":{}}}`),
 		)
 	}
@@ -179,6 +180,7 @@ func TestResProperty(t *testing.T) {
 			GetPropertyVersionResources("prp_0", 1, papi.VersionStatusDeactivated, papi.VersionStatusInactive),
 			SetHostnames("prp_0", 1, "to.test.domain"),
 			AdvanceVersion("prp_0", 1, 2),
+			GetPropertyVersionResources("prp_0", 2, papi.VersionStatusDeactivated, papi.VersionStatusInactive),
 			SetHostnames("prp_0", 2, "to2.test.domain"),
 		),
 		Steps: func(State *TestState, FixturePath string) []resource.TestStep {
@@ -209,6 +211,7 @@ func TestResProperty(t *testing.T) {
 			GetPropertyVersionResources("prp_0", 1, papi.VersionStatusInactive, papi.VersionStatusDeactivated),
 			SetHostnames("prp_0", 1, "to.test.domain"),
 			AdvanceVersion("prp_0", 1, 2),
+			GetPropertyVersionResources("prp_0", 2, papi.VersionStatusInactive, papi.VersionStatusDeactivated),
 			SetHostnames("prp_0", 2, "to2.test.domain"),
 		),
 		Steps: func(State *TestState, FixturePath string) []resource.TestStep {
@@ -239,6 +242,7 @@ func TestResProperty(t *testing.T) {
 			GetPropertyVersionResources("prp_0", 1, papi.VersionStatusActive, papi.VersionStatusInactive),
 			SetHostnames("prp_0", 1, "to.test.domain"),
 			AdvanceVersion("prp_0", 1, 2),
+			GetPropertyVersionResources("prp_0", 2, papi.VersionStatusActive, papi.VersionStatusInactive),
 			SetHostnames("prp_0", 2, "to2.test.domain"),
 		),
 		Steps: func(State *TestState, FixturePath string) []resource.TestStep {
@@ -269,6 +273,7 @@ func TestResProperty(t *testing.T) {
 			GetPropertyVersionResources("prp_0", 1, papi.VersionStatusInactive, papi.VersionStatusActive),
 			SetHostnames("prp_0", 1, "to.test.domain"),
 			AdvanceVersion("prp_0", 1, 2),
+			GetPropertyVersionResources("prp_0", 2, papi.VersionStatusInactive, papi.VersionStatusActive),
 			SetHostnames("prp_0", 2, "to2.test.domain"),
 		),
 		Steps: func(State *TestState, FixturePath string) []resource.TestStep {
@@ -321,6 +326,7 @@ func TestResProperty(t *testing.T) {
 		Name: "No diff found in update",
 		ClientSetup: ComposeBehaviors(
 			PropertyLifecycle("test property", "prp_0", "grp_0"),
+			GetPropertyVersionResources("prp_0", 1, papi.VersionStatusInactive, papi.VersionStatusInactive),
 			SetHostnames("prp_0", 1, "to.test.domain"),
 		),
 		Steps: func(State *TestState, FixturePath string) []resource.TestStep {
@@ -433,6 +439,7 @@ func TestResProperty(t *testing.T) {
 
 			setup := ComposeBehaviors(
 				PropertyLifecycle("test property", "prp_0", "grp_0"),
+				GetPropertyVersionResources("prp_0", 1, papi.VersionStatusInactive, papi.VersionStatusInactive),
 				ImportProperty("prp_0"),
 				SetHostnames("prp_0", 1, "to.test.domain"),
 			)
@@ -554,7 +561,9 @@ func TestResProperty(t *testing.T) {
 
 			setup := ComposeBehaviors(
 				PropertyLifecycle("test property", "prp_0", "grp_0"),
+				GetPropertyVersionResources("prp_0", 1, papi.VersionStatusInactive, papi.VersionStatusInactive),
 				PropertyLifecycle("renamed property", "prp_1", "grp_0"),
+				GetPropertyVersionResources("prp_1", 1, papi.VersionStatusInactive, papi.VersionStatusInactive),
 				SetHostnames("prp_0", 1, "to.test.domain"),
 				SetHostnames("prp_1", 1, "to2.test.domain"),
 			)
@@ -592,6 +601,7 @@ func TestResProperty(t *testing.T) {
 				CreateProperty("test property", "prp_0"),
 				GetProperty("prp_0"),
 				GetVersionResources("prp_0", 1),
+				GetPropertyVersionResources("prp_0", 1, "ctr_0", "grp_0"),
 				SetHostnames("prp_0", 1, "to.test.domain"),
 			)
 			setup(&TestState{Client: client})