+++ pkg/providers/property/testdata/TestPropertyActivation/ok/resource_property_activation_update.tf
@@ -6,4 +6,5 @@ resource "akamai_property_activation" "test" {
   property_id = "test"
   contact = ["user@example.com"]
   version = 2
+  auto_acknowledge_rule_warnings = true
 }